requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
tensorflow==2.12.0
numpy>=1.23.0
opencv-python-headless>=4.7.0
Pillow>=9.0.0
pydantic>=2.0.0
